Allow bills to be viewed or download in the browser
This implements several improvements:

1. Logos specified by GP must be absolute paths or are interpreted relative to the app directory
2. The logo is only loaded from the classpath loader when needed
3. The generated PDF is handled entirely in memory, rather than being written to disk and then loaded into memory

On 1, there's still an issue that absolute paths can specify any file on the
system readable by the Tomcat user, but that's a bigger problem.

On 3, while it may seem like flushing to disk is a good thing in case the file
is large, since we were already reading it to memory to write it to HTTP, the
disk-write is pure overhead.

In the future we should consider:

1. Caching bills to disk so they are only updated when the bill has changed
2. Restricting logos to being only load from the OpenMRS Application Directory